Updated ft_strlcpy (Changed 'int i' to 'size_t i')

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: hhamza <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 09:44:38 by hhamza            #+#    #+#             */
-/*   Updated: 2021/11/03 11:23:51 by hhamza           ###   ########.fr       */
+/*   Updated: 2021/11/03 15:02:55 by hhamza           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((dstsize--) - 1)
